NIFI-5248 Updated PutElasticsearchRecord to only do index operations.
